\name{Bond_PriceDistribution}
\alias{Bond_PriceDistribution}
\alias{Bond_PriceDistribution,Scenarios-method}
\title{Bond_PriceDistribution method}
\arguments{
  \item{t}{Date of pricing (has to be an integer)}

  \item{T}{Date of maturity}

  \item{nCoupons}{Number of coupons}

  \item{couponsRate}{Rate of coupons}
}
\description{
  Get a distribution for bond price.
}

