\name{CBond_PriceDistribution}
\alias{CBond_PriceDistribution}
\alias{CBond_PriceDistribution,Scenarios-method}
\title{CBond_PriceDistribution method}
\arguments{
  \item{t}{Date of pricing (has to be an integer)}

  \item{T}{Date of maturity}

  \item{nCoupons}{Number of coupons}

  \item{couponsRate}{Rate of coupons}

  \item{omega}{Recoverables in case of default}
}
\description{
  Get a distribution for corporate bond price.
}

