\name{CDSPremium_PriceDistribution}
\alias{CDSPremium_PriceDistribution}
\alias{CDSPremium_PriceDistribution,Scenarios-method}
\title{CDSPremium_PriceDistribution}
\arguments{
  \item{t}{Date of pricing (has to be an integer)}

  \item{T}{Date of maturity for the option}

  \item{omega}{Recoverables in case of default}
}
\description{
  Proceed to the projection using the parameters that were
  previously set into the Scenarios objet.
}

