\name{ESG-package}
\alias{ESG-package}
\alias{ESG}
\docType{package}
\title{
ESG - Economic Scenario Generator
}
\description{
Risk neutral Economic Scenario Generator.
}
\details{
\tabular{ll}{
Package: \tab ESG\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2020-11-29\cr
License: \tab GNU\cr
Depends: \tab methods\cr
}

The package is build around the "Scenarios" object.

Use the dedicated methods to set all the needed parameters then use the customPathsGeneration() method to proceed to the asset projection.

}
\keyword{ package }

