\name{EuroCall_Stock_PriceDistribution}
\alias{EuroCall_Stock_PriceDistribution}
\alias{EuroCall_Stock_PriceDistribution,Scenarios-method}
\title{EuroCall_Stock_PriceDistribution method}
\arguments{
  \item{t}{Date of pricing (has to be an integer)}

  \item{T}{Date of maturity}

  \item{Strike}{Strike of the option}
}
\description{
  Get a distribution for EuroCall UL price.
}

