\name{EuroPut_Stock_PriceDistribution}
\alias{EuroPut_Stock_PriceDistribution}
\alias{EuroPut_Stock_PriceDistribution,Scenarios-method}
\title{EuroPut_Stock_PriceDistribution method}
\arguments{
  \item{t}{Date of pricing (has to be an integer)}

  \item{T}{Date of maturity}

  \item{Strike}{Strike of the option}
}
\description{
  Get a distribution for EuroPut UL price.
}

