\name{EuroPut_ZC_PriceDistribution}
\alias{EuroPut_ZC_PriceDistribution}
\alias{EuroPut_ZC_PriceDistribution,Scenarios-method}
\title{EuroPut_ZC_PriceDistribution method}
\arguments{
  \item{t}{Date of pricing (has to be an integer)}

  \item{T}{Date of maturity}

  \item{s}{Date of maturity for the underlying}

  \item{Strike}{Strike of the option}
}
\description{
  Get a distribution for EuroPut ZC price.
}

