\name{Scenarios}
\alias{Scenarios}
\alias{Scenarios-class}
\title{Scenarios class}
\description{
  This is the main class of the package. It has several
  method to read and write the parameters.
}
\details{
  \describe{ \item{ParamsScenarios}{A ParamsScenarios
  object containing the risk parameters}

  \item{ForwardRates}{The forward rates}

  \item{ZCRates}{Volatility for rates in vasicek model}

  \item{shortRatePaths}{The short rate generated paths}

  \item{stockPaths}{The stock generated paths}

  \item{realEstatePaths}{The real estate generated paths}

  \item{liquiditySpreadPaths}{The liquidity spread
  generated paths}

  \item{liquiditySpreadPaths}{The liquidity spread
  generated paths}

  \item{defaultSpreadPaths}{The default spread generated
  paths} }
}

