\name{setRiskParamsScenariosdefSpr}
\alias{setRiskParamsScenariosdefSpr}
\alias{setRiskParamsScenariosdefSpr,Scenarios-method}
\title{setRiskParamsScenariosdefSpr method}
\arguments{
  \item{volDefault}{Volatility for LMN model}

  \item{defaultSpread0}{Initial default spread for LMN
  model}

  \item{alpha}{alpha for LMN model}

  \item{beta}{beta Volatility for LMN model}
}
\description{
  Set risk parameters related to default spread in a
  Scenarios object (these parameters are contained in a
  [\code{\linkS4class{ParamsScenarios}}] sub-object)
}

