\name{setRiskParamsScenariosliqSpr}
\alias{setRiskParamsScenariosliqSpr}
\alias{setRiskParamsScenariosliqSpr,Scenarios-method}
\title{setRiskParamsScenariosliqSpr method}
\arguments{
  \item{eta}{eta Volatility for LMN model}

  \item{liquiditySpread0}{Initial liquidity for LMN model}
}
\description{
  Set risk parameters related to the spread in a Scenarios
  object (these parameters are contained in a
  [\code{\linkS4class{ParamsScenarios}}] sub-object)
}

