\name{setRiskParamsScenariosRE}
\alias{setRiskParamsScenariosRE}
\alias{setRiskParamsScenariosRE,Scenarios-method}
\title{setRiskParamsScenariosRE method}
\arguments{
  \item{vol}{Volatility for rates in vasicek model}

  \item{k}{k for rates in vasicek model}

  \item{volRealEstate}{Volatility for real estate in Black
  & Scholes model}

  \item{realEstate0}{Real estate initial value}
}
\description{
  Set risk parameters related to real estates in a
  Scenarios object (these parameters are contained in a
  [\code{\linkS4class{ParamsScenarios}}] sub-object)
}

