\name{setRiskParamsScenariosrt}
\alias{setRiskParamsScenariosrt}
\alias{setRiskParamsScenariosrt,Scenarios-method}
\title{setRiskParamsScenariosrt method}
\arguments{
  \item{vol}{Volatility for rates in vasicek model}

  \item{k}{k for rates in vasicek model}
}
\description{
  Set risk parameters related to short rates in a Scenarios
  object (these parameters are contained in a
  [\code{\linkS4class{ParamsScenarios}}] sub-object)
}

