\name{ZCBond_PriceDistribution}
\alias{ZCBond_PriceDistribution}
\alias{ZCBond_PriceDistribution,Scenarios-method}
\title{ZCBond_PriceDistribution method}
\arguments{
  \item{t}{Date of pricing (has to be an integer)}

  \item{T}{Date of maturity}
}
\description{
  Get a distribution for ZC bond price.
}

